+++ tests/integration/error-summary.test.js
@@ -10,12 +10,12 @@ describe('Error summary module', () => {
   describe('if an error summary exists', () => {
     beforeEach(() => {
       document.body.innerHTML =
-        '<div role="alert" tabindex="-1" class="nhsuk-error-summary"></div>';
+        '<div role="alert" tabindex="-1" class="ofh-error-summary"></div>';
       ErrorSummary();
     });
 
     it('should focus on the error summary', () => {
-      const errorSummary = document.querySelector('.nhsuk-error-summary');
+      const errorSummary = document.querySelector('.ofh-error-summary');
 
       expect(document.activeElement).toBe(errorSummary);
     });
@@ -23,7 +23,7 @@ describe('Error summary module', () => {
     it('should focus only on the first instance of the error summary', () => {
       const div = document.createElement('div');
       div.innerHTML =
-        '<div role="alert" tabindex="-1" class="nhsuk-error-summary"></div>';
+        '<div role="alert" tabindex="-1" class="ofh-error-summary"></div>';
 
       document.body.appendChild(div.firstChild);