+++ tests/integration/skip-link.test.js
@@ -1,7 +1,7 @@
 import SkipLink from '../../packages/components/skip-link/skip-link';
 
 // Mock HTML
-const skipLinkHtml = '<a class="nhsuk-skip-link" href="#maincontent">Skip to main content</a>';
+const skipLinkHtml = '<a class="ofh-skip-link" href="#maincontent">Skip to main content</a>';
 const headingHtml = '<h1>Test Heading</h1>';
 
 // DOM Elements to be set
@@ -12,11 +12,11 @@ let heading;
 const initTest = (html = '') => {
   document.body.innerHTML = html;
   heading = document.querySelector('h1');
-  skipLink = document.querySelector('.nhsuk-skip-link');
+  skipLink = document.querySelector('.ofh-skip-link');
   SkipLink();
 };
 
-describe('NHS.UK skiplink', () => {
+describe('Our Future Health skiplink', () => {
   describe('Does not throw an error', () => {
     it('if no skiplink exists', () => {
       initTest(headingHtml);