+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Whilst in the alpha phase, we don't yet adhere to [Semantic Versioning](https://
 - Output asset files renamed to `ofh-design-system-toolkit.*`.
 - All favicons and logos updated to latest Our Future Health branding.
 - Renamed the JS and SCSS source entry files to `ofh.*` (note: the build still outputs the compiled files as `ofh-design-system-toolkit.*`).
+- Renamed the prefix for *all* CSS class names from `nhsuk-` to `ofh-`.
 
 ## [v2.0.0-alpha.0] - 2022-07-20